Consolidate flexes for information section body
Adjusted the div containers so that instead of having two horizontal
flexes for the text and the boxes, there is one that contains both
within the main body. Allowed for proper alignment of text under the
boxes.

Additionally, size changes were made to the boxes and the image for
better proportionality. Button border-radius and padding were also
adjusted to better meet the project specifications.